--- com.hobbing.reservation-pay/reservation_pay/src/main/java/com/hobbing/reservation_pay/presentation/dto/PostReservationResBody.java
@@ -1,4 +1,0 @@
-package com.hobbing.reservation_pay.presentation.dto;
-
-public class PostReservationResBody {
-}
