+++ com.hobbing.reservation-pay/reservation_pay/src/main/java/com/hobbing/reservation_pay/presentation/dto/payment/GetPaymentResBody.java
@@ -1,20 +1,22 @@
-package com.hobbing.reservation_pay.presentation.dto;
-
+package com.hobbing.reservation_pay.presentation.dto.payment;
 
 import com.hobbing.reservation_pay.domain.model.Payment;
+import com.hobbing.reservation_pay.domain.model.status_enum.PaymentStatus;
 import lombok.Builder;
 import lombok.Value;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
 
+
 @Value
 @Builder
-public class StudentSearchedPaymentRes {
+public class GetPaymentResBody {
 
-    UUID id;
     UUID couponId;
     String couponName;
+    String receipt;
+    PaymentStatus status;
     int payedPrice;
     String transactionPgToken;
     int discountedPrice;
@@ -23,12 +25,13 @@ public class StudentSearchedPaymentRes {
     LocalDateTime updatedAt;
     UUID updatedBy;
 
-    public static StudentSearchedPaymentRes from(Payment payment) {
 
-        return StudentSearchedPaymentRes.builder()
-                .id(payment.getId())
+    public static GetPaymentResBody from(Payment payment) {
+        return GetPaymentResBody.builder()
                 .couponId(payment.getCouponId())
                 .couponName(payment.getCouponName())
+                .receipt(payment.getReceipt())
+                .status(payment.getStatus())
                 .payedPrice(payment.getPayedPrice())
                 .transactionPgToken(payment.getTransactionPgToken())
                 .discountedPrice(payment.getDiscountedPrice())
@@ -39,3 +42,4 @@ public class StudentSearchedPaymentRes {
                 .build();
     }
 }
+
