+++ com.hobbing.reservation-pay/reservation_pay/src/main/java/com/hobbing/reservation_pay/presentation/dto/payment/PostPaymentReqBody.java
@@ -1,4 +1,4 @@
-package com.hobbing.reservation_pay.presentation.dto;
+package com.hobbing.reservation_pay.presentation.dto.payment;
 
 
 import com.fasterxml.jackson.annotation.JsonCreator;
