+++ com.hobbing.reservation-pay/reservation_pay/src/main/java/com/hobbing/reservation_pay/presentation/dto/reservation/PostReservationResBody.java
@@ -1,0 +1,4 @@
+package com.hobbing.reservation_pay.presentation.dto.reservation;
+
+public class PostReservationResBody {
+}
