+++ com.hobbing.reservation-pay/reservation_pay/src/main/java/com/hobbing/reservation_pay/presentation/dto/reservation/SearchReservationsReqParams.java
@@ -1,7 +1,8 @@
-package com.hobbing.reservation_pay.presentation.dto;
+package com.hobbing.reservation_pay.presentation.dto.reservation;
 
 
 import com.hobbing.reservation_pay.application.dto.SearchReservationsDto;
+import com.hobbing.reservation_pay.presentation.dto.PageInfo;
 import jakarta.validation.constraints.NotNull;
 import lombok.Value;
 import org.springframework.format.annotation.DateTimeFormat;
