+++ com.hobbing.reservation-pay/reservation_pay/src/main/java/com/hobbing/reservation_pay/presentation/dto/settlement/GetSettlementResBody.java
@@ -1,4 +1,4 @@
-package com.hobbing.reservation_pay.presentation.dto;
+package com.hobbing.reservation_pay.presentation.dto.settlement;
 
 
 import com.hobbing.reservation_pay.domain.model.Settlement;
