+++ com.hobbing.reservation-pay/reservation_pay/src/main/java/com/hobbing/reservation_pay/presentation/dto/settlement/SearchSettlementsReqParams.java
@@ -1,7 +1,8 @@
-package com.hobbing.reservation_pay.presentation.dto;
+package com.hobbing.reservation_pay.presentation.dto.settlement;
 
 
 import com.hobbing.reservation_pay.application.dto.SearchSettlementsDto;
+import com.hobbing.reservation_pay.presentation.dto.PageInfo;
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
 import lombok.Value;
